requirements.txt
altair==5.5.0
catboost==1.2.7
lightgbm==4.6.0
numpy==1.26.4
pandas==2.2.3
Pillow==11.1.0
plotly==6.0.0
scikit_learn==1.6.1
st_theme==1.2.3
streamlit==1.43.0
xgboost==3.0.0
